% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc_smc}
\alias{abc_smc}
\title{abc function}
\usage{
abc_smc(
  ref_tree,
  min_lin,
  max_lin,
  statistics,
  simulation_function,
  init_epsilon_value,
  prior_generating_function,
  prior_density_function,
  prior_means,
  number_of_particles = 1000,
  sigma = 0.05,
  stop_rate = 1e-05,
  num_iterations = 50,
  num_threads = 1
)
}
\arguments{
\item{statistics}{A vector containing functions that take a tree
as an argument and return a single scalar value (the statistic).}

\item{simulation_function}{A function that implements the
diversification model and returns an object of class \code{"phylo"}.}

\item{prior_generating_function}{Function to generate parameters
from the prior distribution of these parameters (e.g. a function returning
lambda and mu in case of the birth-death model)}

\item{prior_density_function}{Function to calculate the prior probability
of a set of parameters.}

\item{number_of_particles}{Number of particles to be used
per iteration of the ABC-SMC algorithm.}

\item{sigma}{Standard deviation of the perturbance distribution
(perturbance distribution is a gaussian with mean 0).}

\item{stop_rate}{If the acceptance rate drops below \code{stopRate},
stop the ABC-SMC algorithm  and assume convergence.}

\item{num_iterations}{num iterations}

\item{num_threads}{num_threads}

\item{tree}{an object of class \code{"phylo"}; the tree upon which we want
to fit our diversification model
@param min_lin min_lin
@param max_lin max_lin}

\item{init_epsilon_values}{A vector containing the initial threshold values
for the summary statistics from the vector \code{statistics}.}
}
\value{
A matrix with \code{n} columns,
  where \code{n} is the number of parameters you are trying to estimate.
}
\description{
abc function
}
\references{
Toni, T., Welch, D., Strelkowa, N., Ipsen, A.,
  & Stumpf, M.P.H. (2009). Approximate Bayesian computation scheme for
  parameter inference and model selection in dynamical systems.
  Journal of the Royal Society Interface, 6(31), 187-202.
}
