\name{branchingPhySim.times}
\alias{branchingPhySim.times}
\title{Internal function}
\description{
 An internal function called by several functions in PhySim}
\note{
branchingPhySim.times is derived from branching.times obtained from old versions of the package APE (<=1.8-4) by Emmanuel Paradis.  
}
\author{Emmanuel Paradis}
}
\keyword{internal}
