% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_ref_table_serial}
\alias{create_ref_table_serial}
\title{simulate many trees drawing from the prior serially}
\usage{
create_ref_table_serial(
  num_repl,
  prior_means,
  crown_age,
  min_lin,
  max_lin,
  num_threads
)
}
\arguments{
\item{num_repl}{a vector that indicates the time points of water level changes}

\item{prior_means}{means of 5 priors}

\item{crown_age}{crown age of the tree to be simulated}

\item{min_lin}{minimum number of lineages}

\item{max_lin}{maximum number of lineages}

\item{num_threads}{number of threads}
}
\value{
list with two entries: 1) list of newick strings, 2) matrix of
used parameter values
}
\description{
simulate many trees drawing from the prior serially
}
