% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_ref_table_tbb_par}
\alias{create_ref_table_tbb_par}
\title{simulate many trees drawing from the prior}
\usage{
create_ref_table_tbb_par(num_repl, prior_means, crown_age, min_lin, max_lin)
}
\arguments{
\item{num_repl}{a vector that indicates the time points of water level changes}

\item{crown_age}{crown age of the tree to be simulated}

\item{min_lin}{minimum number of lineages}

\item{max_lin}{maximum number of lineages}
}
\value{
list with two entries: 1) list of newick strings, 2) matrix of
used parameter values
}
\description{
simulate many trees drawing from the prior
}
