% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dd_R.R
\name{dd_sim_R}
\alias{dd_sim_R}
\title{analagous function to the DDD function, with intermediate size checking}
\usage{
dd_sim_R(pars, age, num_lin = -1)
}
\arguments{
\item{pars}{pars}

\item{age}{age}

\item{num_lin}{num_lin}
}
\value{
list
}
\description{
analagous function to the DDD function, with intermediate size checking
}
