% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_from_prior_rcpp}
\alias{draw_from_prior_rcpp}
\title{function to generate 5 parameters from exponential prior, given mean
values}
\usage{
draw_from_prior_rcpp(prior_means)
}
\arguments{
\item{prior_means}{vector of prior means}
}
\value{
vector of values
}
\description{
function to generate 5 parameters from exponential prior, given mean
values
}
