\name{dropPhySim.tip}
\alias{dropPhySim.tip}
\title{Internal function}
\description{
 An internal function called by several functions in PhySim}
\note{
dropPhySim.tip is derived from drop.tip obtained from old versions of the package APE (<=1.8-4) by Emmanuel Paradis.  
}
\author{Emmanuel Paradis}
}
\keyword{internal}
