% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_sim_rcpp.R
\name{pbd_sim_conditional_rcpp}
\alias{pbd_sim_conditional_rcpp}
\title{Function to simulate the protracted speciation process}
\usage{
pbd_sim_conditional_rcpp(
  pars,
  age,
  min_species = 300,
  max_species = 300,
  num_tries = 100
)
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr \code{pars[1]} corresponds to b_1,
the speciation-initiation rate of good species \cr \code{pars[2]}
corresponds to la_1, the speciation-completion rate \cr \code{pars[3]}
corresponds to b_2, the speciation-initiation rate of incipient species \cr
\code{pars[4]} corresponds to mu_1, the extinction rate of good species \cr
\code{pars[5]} corresponds to mu_2, the extinction rate of incipient species
\cr}

\item{age}{Sets the crown age for the simulation}

\item{min_species}{minimum number of species in tree}

\item{max_species}{maximum number of species in tree}

\item{num_tries}{number of times to re-start the simulation in case of
extinction (all species dead) or overflow (if upper_species_limit is
exceeded)}
}
\value{
phy object in case of success, or a string ("extinction" or
"overflow",
if no tree could be generated within num_tries tries).
}
\description{
Function to simulate the protracted speciation process
}
\author{
Thijs Janzen
}
\keyword{models}
