% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perform_abc_rcpp}
\alias{perform_abc_rcpp}
\title{function to do abc using rcpp}
\usage{
perform_abc_rcpp(
  num_particles,
  num_iterations,
  crown_age,
  min_lin,
  max_lin,
  lambdas,
  s,
  obs_gamma,
  obs_colless,
  obs_num_lin
)
}
\arguments{
\item{num_particles}{number of particles}

\item{num_iterations}{number of iterations}

\item{crown_age}{crown age}

\item{min_lin}{minimum number of lineages from the prior}

\item{max_lin}{maximum number oflineages from the prior}

\item{lambdas}{vector of lambdas for exponential priors (5)}

\item{s}{perturbation standard deviation. Perturbations are made on a log scale,
e.g. new_param = exp( log(s) + N(0, s))}

\item{obs_gamma}{observed gamma value to fit on}

\item{obs_colless}{observed colless value to fit on}

\item{obs_num_lin}{observed number of lineages to fit on}
}
\description{
function to do abc using rcpp
}
