% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prior_dens_rcpp}
\alias{prior_dens_rcpp}
\title{function to calculate density of parameters}
\usage{
prior_dens_rcpp(prior_means, x)
}
\arguments{
\item{prior_means}{vector of prior means}

\item{x}{vector of x values for which to calculate the density}
}
\value{
vector of values
}
\description{
function to calculate density of parameters
}
