% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bisse.R
\name{sim_bisse}
\alias{sim_bisse}
\title{function so simulate bisse model}
\usage{
sim_bisse(
  pars,
  crown_age = 10,
  num_species = 300,
  init_state = -1,
  verbose = FALSE,
  max_tries = 1e+06
)
}
\arguments{
\item{pars}{parameters as vector {l0, l1, m0, m1, q01, q10}}

\item{crown_age}{crown age}

\item{num_species}{maximum number of species allowed}

\item{init_state}{initial state, if set to -1, 0 or 1 is randomly drawn}

\item{verbose}{generate verbose output}

\item{max_tries}{maximum number of tries to generate tree}

\item{use_sim1}{use sim1 method}
}
\value{
list
}
\description{
function so simulate bisse model
}
