% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dd.R
\name{sim_ddd}
\alias{sim_ddd}
\title{function to simulate the linear version of the Diversity-Dependent-Diversification
model}
\usage{
sim_ddd(
  lambda,
  mu,
  K,
  max_t = -1,
  num_species = -1,
  return_fossils = FALSE,
  seed = -1
)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{K}{level of diversity at which the speciation rate equals the extinction
rate}

\item{max_t}{crown age of the tree}

\item{num_species}{number of species of the tree}

\item{return_fossils}{should extinct species be returned in the tree?}

\item{seed}{seed of the random number generator}
}
\value{
phylo object
}
\description{
Simulations are performed either conditional on crown age, on the number of
extant lineages, or on both, depending on which of max_t and num_species has
a value larger than zero.
}
