% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pbd.R
\name{sim_pbd}
\alias{sim_pbd}
\title{function to simulate the linear version of the Diversity-Dependent-Diversification
model}
\usage{
sim_pbd(
  lambda0,
  mu0,
  lambda1,
  mu1,
  completion_rate,
  max_t = -1,
  num_species = -1,
  max_tries = 1e+06,
  seed = -1
)
}
\arguments{
\item{lambda0}{speciation rate}

\item{mu0}{extinction rate}

\item{mu1}{extinction rate2}

\item{completion_rate}{completion rate}

\item{max_t}{crown age of the tree}

\item{num_species}{number of species of the tree}

\item{seed}{seed of the random number generator}

\item{labmda1}{speciation rate2}
}
\value{
phylo object
}
\description{
Simulations are performed either conditional on crown age, on the number of
extant lineages, or on both, depending on which of max_t and num_species has
a value larger than zero.
}
