% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pbd_cpp}
\alias{sim_pbd_cpp}
\title{rcpp implementation}
\usage{
sim_pbd_cpp(
  la0,
  mu0,
  la1,
  mu1,
  trans_rate,
  max_t,
  max_num_species,
  num_tries = 100L
)
}
\arguments{
\item{la0}{l0}

\item{mu0}{xx}

\item{la1}{xx}

\item{mu1}{xx}

\item{trans_rate}{xx}

\item{max_t}{max_t}

\item{max_num_species}{max_num_species}
}
\description{
rcpp implementation
}
